# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API Specification
  description: |-
      This OpenAPI document describes a set of interfaces to allow a WASAPhoto user to:
      - See a stream of posts
      - Add and remove likes from posts
      - Interact with the posts with comments
      - Allow the author to remove comments
      - Ban and unban other users
      - See their personal profile
      - Set the username
      - Add and remove personal posts
      - Follow and unfollow other users
      - Search for other users
        
  version: "1.0.0"
  contact:
    name: Ionut Radu Barbalata
    email: barbalata.2002688@studenti.uniroma1.it
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    UniversallyUniqueIdentifiers:
      title: Universally Unique Identifier
      description: |
        A random based UUID used to identify UserID, PhotoID, and more.
      type: string
      pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
      example: "550e8400-e29b-41d4-a716-446655440001"
      minLength: 36
      maxLength: 36
    Username:
      title: Username
      description: |
        Represent a possible name of a User.
      type: object 
      properties:
        username_string:
          description: string of the object
          type: string
          pattern: '^[a-zA-Z0-9_.]{3,20}$'
          example: "Ionut_02"
          minLength: 3
          maxLength: 20
      example:
        username_string: "Jok3r"
...
    PostPrototype:
      title: Prototype post
      type: object
      description: | 
        Contain al the information needed to initialize
        a post
      properties:
        file:
          description: A binary file that represent the image
          type: string
          format: binary
          minLength: 0
          maxLength: 30000000 # 30 MB
          example: 0101010110
        description:
          description: text related to the post
          type: string
          pattern: (.|\n)*
          minLength: 0
          maxLength: 2200
          example: it's raining
    Post:
      title: Post
      type: object
      description: An object that rapresent a post
      properties:
        upload_time:
          description: Time when was uploaded the post
          type: string
          format: date-time
          example: "2017-07-21T17:32:28Z"
        author:
          description: string of the object
          type: string
          pattern: '^[a-zA-Z0-9_.]{3,20}$'
          example: "Ionut_02"
          minLength: 3
          maxLength: 20
        num_likes:
          description: number of likes of the photo
          type: integer
          minimum: 0
        num_comments:
          description: number of comments of the photo
          type: integer
          minimum: 0
        liked_photo:
          description: if True then you liked it
          type: boolean
        photo_id:
          $ref: "#/components/schemas/UniversallyUniqueIdentifiers"
        photo_file:
          description: A binary file that represent the image
          type: string
          format: binary
          minLength: 0
          maxLength: 30000000 # 30 MB
        description: 
          description: text related to the post
          type: string
          pattern: (.|\n)*
          minLength: 0
          maxLength: 2200

      example:
        upload_time: "2023-11-10T10:15:00Z"
        author: "Alice123"
        num_likes: 20
        num_comments: 5
        liked_photo: True
        photo_id: "550e8400-e29b-41d4-a716-446655440001"
        photo_file: "000010101010101"
        description: "it's raining"
    PostCollection:
      title: Posts
      description: An array of posts
      type: object
      properties:
        posts:
          description: posts 
          type: array
          minItems: 0
          maxItems: 200
          items:
            $ref: "#/components/schemas/Post"
      example:
        posts:
          - upload_time: "2023-11-10T10:15:00Z"
            author: "Alice123"
            num_likes: 20
            num_comments: 5
            liked_photo: True
            photo_id: "550e8400-e29b-41d4-a716-446655440001"
            photo_file: "000010101010101"
            description: "it's raining"
          - upload_time: "2023-10-10T10:15:00Z"
            author: "Alice123"
            num_likes: 20
            num_comments: 5
            liked_photo: True
            photo_id: "550e8400-e29b-41d4-a716-446655440001"
            photo_file: "000010101010101"
            description: "it's raining"
...
    Response:
      title: Rersonse
      description: A response message, it contains a code and a message
      type: object
      properties:
        response:
          description: The error message
          type: string
          pattern: (.|\n)*
          minLength: 1
          maxLength: 150
          example: Bad request

        code:
          description: code error
          type: integer
          minimum: 200
          maximum: 599
          example: 400
...
paths:
...
  /users/{username}/profile/posts/:
    parameters:
      - name: username
        description: The username of the one to get the posts
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Username"
    get:
      tags: ["posts"]
      operationId: getUserPosts
      summary: Get the user Posts
      description: | 
        Get the posts of the user and verify if the
          user of this session is banned or not
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Succesfully get the user's posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostCollection"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
    post:
      tags: ["posts"]
      operationId: uploadPhoto
      summary: Create a new Post
      description: |
        Insert in the username profile a new Post with can be
          available online
      security:
        - bearerAuth: []
      requestBody:
        content: 
          application/json:
            schema: 
              $ref: "#/components/schemas/PostPrototype"
      responses:
        "201": 
          description: Succesfully uploaded a new post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /users/{username}/profile/posts/{post_id}:
    parameters:
      - name: username
        description: The user's username
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Username"
      - name: post_id
        description: The UUID of the post
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UniversallyUniqueIdentifiers"
    
    get:
      tags: ["posts"]
      operationId: getPost
      summary: Get the post with the postId
      description: | 
        Get the post of the user and verify if the
          user of this session is banned or not
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Succesfully get the user's post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
    delete:
      tags: ["posts"]
      operationId: deletePhoto
      summary: Delete the post with the postId
      description: | 
        Delete the post of the user and verify if the
          user of this session is the profile owner
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Succesfully deleted the user's post
        "403": 
          description: | 
            Insufficient permissions to delete the resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
...